Extracts ChatForm component
